+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 2019 Sep 25 [in progress]
 
 * Added NaN checks in multimeter when accumulating (Issue #2).
+* Added configurable stream buffer duration (was fixed at 30 seconds).
 
 
 ## 0.6.3